Update int to float for player and point classes

+++ src/dev/firstPerson/Camera.java
@@ -26,7 +26,7 @@ public class Camera {
 
     }
 
-    private void drawSky(int direction, Bitmap skybox, Bitmap light) {
+    private void drawSky(float direction, Bitmap skybox, Bitmap light) {
 
     }
 }
